Extract variables tree data transformation function new module
